// web/src/api/generated/models/AttachmentReadRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AttachmentReadRequest
 */
export interface AttachmentReadRequest {
    /**
     * 
     * @type {string}
     * @memberof AttachmentReadRequest
     */
    parentId?: string | null;
    /**
     * 
     * @type {AttachmentReadRequestParentTypeEnum}
     * @memberof AttachmentReadRequest
     */
    parentType?: AttachmentReadRequestParentTypeEnum;
    /**
     * Server-defined content variant. Request thumbnail only when hasThumbnail is true.
     * @type {AttachmentReadRequestVariantEnum}
     * @memberof AttachmentReadRequest
     */
    variant?: AttachmentReadRequestVariantEnum;
}


/**
 * @export
 */
export const AttachmentReadRequestParentTypeEnum = {
    MEMORY: 'MEMORY',
    HEART_MOMENT: 'HEART_MOMENT',
    RELATED_PERSON: 'RELATED_PERSON',
    NONE: 'NONE'
} as const;
export type AttachmentReadRequestParentTypeEnum = typeof AttachmentReadRequestParentTypeEnum[keyof typeof AttachmentReadRequestParentTypeEnum];

/**
 * @export
 */
export const AttachmentReadRequestVariantEnum = {
    original: 'original',
    thumbnail: 'thumbnail'
} as const;
export type AttachmentReadRequestVariantEnum = typeof AttachmentReadRequestVariantEnum[keyof typeof AttachmentReadRequestVariantEnum];


/**
 * Check if a given object implements the AttachmentReadRequest interface.
 */
export function instanceOfAttachmentReadRequest(value: object): value is AttachmentReadRequest {
    return true;
}

export function AttachmentReadRequestFromJSON(json: any): AttachmentReadRequest {
    return AttachmentReadRequestFromJSONTyped(json, false);
}

export function AttachmentReadRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AttachmentReadRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'parentId': json['parentId'] === undefined ? undefined : json['parentId'] === null ? null : json['parentId'],
        'parentType': json['parentType'] == null ? undefined : json['parentType'],
        'variant': json['variant'] == null ? undefined : json['variant'],
    };
}

export function AttachmentReadRequestToJSON(json: any): AttachmentReadRequest {
    return AttachmentReadRequestToJSONTyped(json, false);
}

export function AttachmentReadRequestToJSONTyped(value?: AttachmentReadRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'parentId': value['parentId'],
        'parentType': value['parentType'],
        'variant': value['variant'],
    };
}